feat: Create a class hierarchy for backend with dependency injection

+++ backend/src/database/services.js
@@ -1,39 +1,62 @@
-const Database = require('better-sqlite3');
-const path = require('path');
-const fs = require('fs');
-const { DB_FOLDER, SCHEMA_PATH } = require('../config/paths');
+import Database from 'better-sqlite3';
 
-const getSafePath = (name) => path.join(DB_FOLDER, `${name.replace(/[^a-z0-9_-]/gi, '_').toLowerCase()}.db`);
+class DBConnector {
 
-const dbService = {
-  getDB(name) {
-    const dbPath = getSafePath(name);
-    if (!fs.existsSync(dbPath)) throw new Error("Compétition introuvable.");
-    return new Database(dbPath);
-  },
+    #paths
+    #fs
 
+    constructor(paths, fs){
+        this.#paths = paths;
+        this.#fs = fs;
+    }
+
+    isDatabaseExists(name) {
+        const dbPath = this.#paths.getSafeDatabasePath(name);
+        return this.#fs.existsSync(dbPath);        
+    }
+    
+    getDatabase(name) {
+        const dbPath = this.#paths.getSafeDatabasePath(name);
+        return new Database(dbPath);
+    }
+
+}
+
+export default class DatabaseService {
+
+    #connector
+    #paths
+    #fs
+
+    constructor(fs, paths,connector = null) {
+        this.#fs        = fs;
+        this.#paths     = paths;
+        this.#connector = connector ? connector : new DBConnector(paths, fs);
+    }
+
+ 
   createCompetition(name) {
-    const safeName = name.replace(/[^a-z0-9_-]/gi, '_').toLowerCase();
-    const dbPath = getSafePath(name);
-    if (fs.existsSync(dbPath)) throw new Error("EXIST_ERROR");
 
-    const db = new Database(dbPath);
-    const schema = fs.readFileSync(SCHEMA_PATH, 'utf8');
+    if (this.#connector.isDatabaseExists(name)) throw new Error("EXIST_ERROR");
+    
+    const db = this.#connector.getDatabase(name);
+    const schema = this.#fs.readFileSync(this.#paths.schemaPath, 'utf8');
     db.exec(schema);
     return safeName;
-  },
+
+  }
 
   syncRacers(competition, racers) {
-    const db = this.getDB(competition);
+
+    if (!this.#connector.isDatabaseExists(competition)) throw new Error("NOT_EXIST_ERROR");
+
+    const db = this.#connector.getDatabase(name);
     const syncTransaction = db.transaction((data) => {
       db.prepare('DELETE FROM racers').run();
       const insert = db.prepare('INSERT INTO racers (id, data) VALUES (?, ?)');
       for (const r of data) insert.run(r.id, JSON.stringify(r));
     });
     syncTransaction(racers);
-  },
+  }
 
-  // ... Ajoutez ici syncRankings, syncRaceInfo, etc.
 };
-
-module.exports = dbService;
